requirements.txt
discord.py
pandas
nunpy
wikipedia